Add format_str parameter for progress bar and mill
progress.Bar, progress.bar, and progress.mill have the new format_str parameter.

Named fields are used to facilitate the change. The respective formats would be
represented using the new syntax below:

    '{label}[{bar}] {completed}/{total} - {eta}'
    '{label} {mill} {completed}/{total}'

str.format() fails gracefully when named fields aren't present in the string,
allowing a given part of the string to be omitted. On the other hand, incorrect
field names cause a KeyError, preventing user error.

Note that when format_str=None, the old templates are used so that any code that
modifies progress.BAR_TEMPLATE, for example, is not affected by the change.

Also, the format_str parameter does not require \r at the end, it is appended
automatically.